+++ aristotle_mdr/managers.py
@@ -8,12 +8,18 @@ from model_utils.managers import InheritanceManager, InheritanceQuerySet
 
 class UUIDManager(models.Manager):
     def create_uuid(self, instance):
+        from aristotle_mdr.models import _concept
+        if type(instance) is _concept or issubclass(type(instance), _concept) :
+            instance = instance.item
         if instance.uuid is not None:
-            return
-        instance.uuid = self.create(
-            app_label=instance._meta.app_label,
-            model_name=instance._meta.model_name,
-        )
+            instance.uuid.app_label=instance._meta.app_label
+            instance.uuid.model_name=instance._meta.model_name
+            instance.uuid.save()
+        else:
+            instance.uuid = self.create(
+                app_label=instance._meta.app_label,
+                model_name=instance._meta.model_name,
+            )
 
 
 class MetadataItemQuerySet(InheritanceQuerySet):