+++ aristotle_mdr/migrations/0024_add_uuid_instances.py
@@ -14,17 +14,19 @@ class Migration(migrations.Migration):
     def dependencies(cls):
         deps = [
             ('aristotle_mdr', '0023_uuid_model'),
-            ('aristotle_mdr_links', '0005_switch_to_concept_relations'),
             ('aristotle_mdr_slots', '0004_switch_to_concept_relations'),
         ]
         from django.conf import settings
+
+        if "aristotle_mdr.contrib.links" in settings.INSTALLED_APPS:
+            deps.append(
+                ('aristotle_mdr_links', '0005_switch_to_concept_relations'),
+            )
 
         if "aristotle_multisite" in settings.INSTALLED_APPS:
             deps.append(
                 ('aristotle_multisite', '0001_initial'),
             )
-
-        return deps
 
     operations = [
         # This is needed as we modify ForeignKeys before they are made during the migration