+++ aristotle_mdr/views/user_pages.py
@@ -96,7 +96,7 @@ def admin_tools(request):
                 (
                     m.model_class(),
                     get_cached_object_count(m),
-                    reverse("admin:%s_%s_changelist" % (m.app_label, m.model))
+                    reverse("browse_concepts", args=[m.app_label, m.model])
                 )
             )
             model_stats[m.app_label] = app_models
@@ -111,7 +111,9 @@ def admin_tools(request):
 
 @login_required
 def admin_stats(request):
-    if not request.user.is_superuser:
+    if request.user.is_anonymous():
+        return redirect(reverse('friendly_login') + '?next=%s' % request.path)
+    elif not request.user.has_perm("aristotle_mdr.access_aristotle_dashboard"):
         raise PermissionDenied
 
     aristotle_apps = fetch_metadata_apps()
@@ -162,7 +164,7 @@ def admin_stats(request):
                         't7': t7_val,
                         't30': t30_val
                     },
-                    reverse("admin:%s_%s_changelist" % (m.app_label, m.model))
+                    reverse("browse_concepts", args=[m.app_label, m.model])
                 )
             )
             model_stats[m.app_label] = app_models